scripts/gen-asset/requirements.txt
google-genai>=0.3.0
pillow>=10.0.0
